+++ multimodal-app/components/predefined/ExternalLink.tsx
