+++ multimodal-app/components/predefined/HapticTab.tsx
