+++ multimodal-app/components/predefined/ThemedView.tsx
